Project/requirements.txt
asgiref==3.5.2
Django==4.0.5
django-filter==23.2
django-phone-field==1.8.1
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
phonenumbers==8.13.19
Pillow==10.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
